moved math function to kernel::common::math
